fix prcess utilization

+++ include/process.h
@@ -8,7 +8,7 @@ It contains relevant attributes as shown below
 */
 class Process {
  public:
-  long CurrentCpuUsage;
+  float CurrentCpuUsage;
   Process(int pid);
   int Pid();                               // TODO: See src/process.cpp
   std::string User();                      // TODO: See src/process.cpp
